Move emit stuff into own file
